Convert to netstandard 2.0
