+++ DatabaseSampleApp.DB/Models/IHasServerId.cs
@@ -1,0 +1,7 @@
+﻿namespace DatabaseSampleApp.DB.Models
+{
+    public interface IHasServerId
+    {
+        int? ServerId { get; }
+    }
+}
